Add authorization to evaluation item

+++ app/Http/Controllers/Admin/EvaluationItemController.php
@@ -4,20 +4,20 @@ namespace App\Http\Controllers\Admin;
 
 use App\EvaluationItem;
 use App\EvaluationScore;
-use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
-use TiMacDonald\Validation\Rule;
 use Illuminate\Support\Facades\Redirect;
 use Illuminate\Support\Arr;
 use Illuminate\Support\Facades\DB;
 use App\Http\Controllers\Project\EvaluationScoreController;
 use App\Http\Controllers\Project\ProjectController;
 use App\Project;
+use App\Http\Requests\EvaluationItemRequest;
 
 class EvaluationItemController extends Controller
 {
     public function index()
     {
+        $this->authorize('viewAny', EvaluationItem::class);
         $evaluationItems = EvaluationItem::sorted()->get();
         $sum = array_sum(Arr::pluck($evaluationItems, 'weight'));
         return view('admin.evaluation_items.index', compact('evaluationItems', 'sum'));
@@ -25,17 +25,17 @@ class EvaluationItemController extends Controller
 
     public function create()
     {
+        $this->authorize('create', EvaluationItem::class);
         $evaluationItem = new EvaluationItem();
         $formAction = route('admin.evaluation_items.store');
         return view('admin.evaluation_items.edit', compact('evaluationItem', 'formAction'));
     }
 
-    public function store(Request $request)
+    public function store(EvaluationItemRequest $request)
     {
-        $values = $this->validateValues($request);
-
-        DB::transaction(function () use ($values) {
-            $evaluationItem = EvaluationItem::create($values);
+        $this->authorize('create', EvaluationItem::class);
+        DB::transaction(function () use ($request) {
+            $evaluationItem = EvaluationItem::create($request->validated());
             foreach (Project::all() as $project) {
                 $evaluationScore = new EvaluationScore([
                     'score' => 1,
@@ -52,16 +52,16 @@ class EvaluationItemController extends Controller
 
     public function edit(EvaluationItem $evaluationItem)
     {
+        $this->authorize('view', $evaluationItem);
         $formAction = route('admin.evaluation_items.update', ['evaluation_item' => $evaluationItem->pid]);
         return view('admin.evaluation_items.edit', compact('evaluationItem', 'formAction'));
     }
 
-    public function update(Request $request, EvaluationItem $evaluationItem)
+    public function update(EvaluationItemRequest $request, EvaluationItem $evaluationItem)
     {
-        $values = $this->validateValues($request);
-
-        DB::transaction(function () use ($evaluationItem, $values) {
-            $evaluationItem->update($values);
+        $this->authorize('update', $evaluationItem);
+        DB::transaction(function () use ($evaluationItem, $request) {
+            $evaluationItem->update($request->validated());
             $this->recalculate();
         });
 
@@ -70,21 +70,12 @@ class EvaluationItemController extends Controller
 
     public function destroy(EvaluationItem $evaluationItem)
     {
+        $this->authorize('delete', $evaluationItem);
         DB::transaction(function () use ($evaluationItem) {
             $evaluationItem->delete();
             $this->recalculate();
         });
         return Redirect::route('admin.evaluation_items.index');
-    }
-
-    private function validateValues(Request $request)
-    {
-        return $request->validate([
-            'name' => Rule::required()->string(1, EvaluationItem::DD_NAME_LENGTH)->get(),
-            'instructions' => Rule::nullable()->string(1, EvaluationItem::DD_INSTRUCTIONS_LENGTH)->get(),
-            'weight' => Rule::required()->integer(1, EvaluationItem::DD_WEIGHT_MAX)->get(),
-            'sort_order' => Rule::required()->integer(0, EvaluationItem::DD_SORT_ORDER_MAX)->get()
-        ]);
     }
 
     private function recalculate()
